% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_PC_data.R
\name{compile_PC_data}
\alias{compile_PC_data}
\title{Compile point count survey data from a directory}
\usage{
compile_PC_data(dir, pattern = ".csv")
}
\arguments{
\item{dir}{directory containing point count survey data downloaded from CADC}

\item{pattern}{regex pattern for identifying files to include (see \code{\link[base]{list.files}})}
}
\value{
Dataframe including all data, along with added columns converting
"Distance Bin" to mindist and maxdist numeric fields.
}
\description{
Compile point count survey data from a directory
}
