% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing.R
\name{fill_missing}
\alias{fill_missing}
\title{Fill missing values}
\usage{
fill_missing(df)
}
\arguments{
\item{df}{dataframe or tibble}
}
\value{
tibble
}
\description{
Internal utility function called by [summarize_capture_stats()] and
[summarize_nethrs_stats()]; not intended to be called directly
}
\details{
Utility function to expand a dataframe to ensure there is a row for
every location, year, month, and/or season present in the data, as well as
every species (in the case of banding data). For example, used by
[summarize_capture_stats()] to ensure that a lack of captures for a given
species in a given time frame (month, season, or year) is represented with a
0 value. For net hours, to ensure each location is represented in each month,
season, or year, even if there was no effort during that time frame. The
zeroes are by default filled into a column labeled "y".
}
\keyword{internal}
