% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BBS_model.R
\name{plot_BBS_model}
\alias{plot_BBS_model}
\title{Plot trend results from BBS-style hierarchical model for point count survey data}
\usage{
plot_BBS_model(df, type = "both")
}
\arguments{
\item{df}{Summary of median, lower, and upper estimates from \code{\link{get_BBS_model_estimates}}}

\item{type}{Type of plot to produce: plots of annual indices of abundance
('index') or plots of the linear trend lines ('trend'), both ('both'). Defaults to 'both'.}
}
\value{
returns ggplot
}
\description{
Produces line and ribbon for trend estimates, and points with error bars for index estimates using \pkg{ggplot2}. Install it with: \code{install.packages("ggplot2")}
}
