% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{resolution}
\alias{resolution}
\title{Suggest resolution}
\usage{
resolution(
  dim1,
  dim2,
  resolution = c(1:50) * 20/max(diff(range(dim1)), diff(range(dim2))),
  plot = TRUE
)
}
\arguments{
\item{dim1}{Numeric vector for the first dimension, such as the first dimension of UMAP or PCA.}

\item{dim2}{Numeric vector for the second dimension.}

\item{resolution}{A numeric vector of resolution values to be tested.}

\item{plot}{Either TRUE or FALSE. If TRUE, the function will directly generate a plot. If FALSE, the function will return the data frame used for plotting.}
}
\value{
If plot=TRUE, a ggplot object. If plot=FALSE, a vector .
}
\description{
Suggest the resolution to be used by SCUBI
}
\details{
This function suggests the resolution to be used in various SCUBI functions. For each resolution, the averaged number of cells within each square is calculated. To determine the suggested resolution, a plot will be generated to study the change the averaged number of cells with resolution, and the suggested resolution is picked at the elbow location.
}
\examples{
resolution(dim1=rnorm(1000),dim2=rnorm(1000))
}
\author{
Wenpin Hou<whou10@jhu.edu>, Zhicheng Ji
}
