% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution_sample_expression.R
\name{resolution_sample_expression}
\alias{resolution_sample_expression}
\title{Suggest resolution for multi-sample regression}
\usage{
resolution_sample_expression(
  dim1,
  dim2,
  sample,
  design,
  resolution = c(1:50) * 20/max(diff(range(dim1)), diff(range(dim2))),
  plot = TRUE
)
}
\arguments{
\item{dim1}{Numeric vector for the first dimension, such as the first dimension of UMAP or PCA.}

\item{dim2}{Numeric vector for the second dimension.}

\item{sample}{Character vector of cell-sample assignment (each cell belong to which sample).}

\item{design}{A numeric data.frame or matrix indicating the covariates of the samples. It should not contain the intercept term. The row names of the matrix should contain the unique values in 'sample' variable.}

\item{resolution}{A numeric vector of resolution values to be tested.}

\item{plot}{Either TRUE or FALSE. If TRUE, the function will directly generate a plot. If FALSE, the function will return the data frame used for plotting.}
}
\value{
If plot=TRUE, a ggplot object. If plot=FALSE, a vector .
}
\description{
Suggest the resolution to be used by SCUBI multi-sample regression mode
}
\details{
This function suggests the resolution to be used in SCUBI function of multi-sample regression. For each resolution, the averaged number of cells within each square is calculated. To determine the suggested resolution, a plot will be generated to study the change the averaged number of cells with resolution, and the suggested resolution is picked at the elbow location.
}
\examples{
resolution_sample_expression(dim1=rnorm(1000),dim2=rnorm(1000),sample=rep(c('a','b'),500),design=data.frame(c(0,1),row.names=c('a','b')))
}
\author{
Wenpin Hou<whou10@jhu.edu>, Zhicheng Ji
}
