% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_band.R
\docType{data}
\name{sample_band}
\alias{sample_band}
\title{Sample banding data}
\format{
## `sample_band`
A data frame with 23,575 rows and 35 columns
}
\usage{
sample_band
}
\description{
A sample of data from allnumb.dbf for the years 2009-2012, just for the
purposes of demonstrating certain functions.
}
\keyword{datasets}
