% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_nethrs.R
\docType{data}
\name{sample_nethrs}
\alias{sample_nethrs}
\title{Sample net hours data}
\format{
## `sample_nethrs`
A data frame with 1,858 rows and 14 columns
}
\usage{
sample_nethrs
}
\description{
A sample of data from allpalonethrs.dbf for the years 2009-2012, just for the
purposes of demonstrating certain functions.
}
\keyword{datasets}
