% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scubi_continuous.R
\name{scubi_continuous}
\alias{scubi_continuous}
\title{SCUBI for continuous variable}
\usage{
scubi_continuous(
  dim1,
  dim2,
  feature,
  resolution = 10 * 20/max(diff(range(dim1)), diff(range(dim2))),
  plot = TRUE,
  palette = rainbow(15)[c(11:1, 15)]
)
}
\arguments{
\item{dim1}{Numeric vector for the first dimension, such as the first dimension of UMAP or PCA.}

\item{dim2}{Numeric vector for the second dimension.}

\item{feature}{Numeric vector of a continuous variable (e.g. mitochondrial read proportion).}

\item{resolution}{Numeric value of resolution. A higher value will result in smaller squares, thus higher resolution.}

\item{plot}{Either TRUE or FALSE. If TRUE, the function will directly generate a plot. If FALSE, the function will return the data frame used for plotting. This is useful for more complicated operations.}

\item{palette}{A vector of color codes for plotting palette.}
}
\value{
If plot=TRUE, a ggplot object. If plot=FALSE, a data.frame with the following columns: xmin,xmax,ymin,ymax of the square, and the averaged expresison value of that square.
}
\description{
SCUBI for plotting a continuous variable
}
\details{
This function partitions the whole coordinate space into multiple non-overlapping squares, and calculates and visualizes the averaged value of a user-defined continuous variable for cells falling in each square It takes as input the coordinates of the cells (e.g. UMAP or PCA) and the values of the continuous variable.
}
\examples{
scubi_continuous(dim1=rnorm(10000),dim2=rnorm(10000),feature=rnorm(10000))
}
\author{
Wenpin Hou<whou10@jhu.edu>, Zhicheng Ji
}
