% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scubi_expression.R
\name{scubi_expression}
\alias{scubi_expression}
\title{SCUBI for gene expression}
\usage{
scubi_expression(
  dim1,
  dim2,
  count,
  gene,
  resolution = 10 * 20/max(diff(range(dim1)), diff(range(dim2))),
  plot = TRUE,
  palette = rainbow(15)[c(11:1, 15)]
)
}
\arguments{
\item{dim1}{Numeric vector for the first dimension, such as the first dimension of UMAP or PCA.}

\item{dim2}{Numeric vector for the second dimension.}

\item{count}{Count matrix for the gene expression data. Each row is a gene and each column is a cell. A sparse matrix is accepted.}

\item{gene}{A character value, a character vector, or a list of gene names. The gene names must exist in the count matrix. For a character value, the expression of that gene will be plotted. For a character vector, the read counts will be summed across these genes. For a list of gene names, each element is a gene or a vector of genes. A subplot will be generated for each element in the list.}

\item{resolution}{Numeric value of resolution. A higher value will result in smaller squares, thus higher resolution.}

\item{plot}{Either TRUE or FALSE. If TRUE, the function will directly generate a plot. If FALSE, the function will return the data frame used for plotting. This is useful for more complicated operations.}

\item{palette}{A vector of color codes for plotting palette.}
}
\value{
If plot=TRUE, a ggplot object. If plot=FALSE, a data.frame with the following columns: xmin,xmax,ymin,ymax of the square, and the averaged expresison value of that square.
}
\description{
SCUBI for plotting the expression of one gene
}
\details{
This function partitions the whole coordinate space into multiple non-overlapping squares, and calculates and visualizes the averaged expression of cells falling in each square. It takes as input the coordinates of the cells (e.g. UMAP or PCA), the gene expression count matrix, and the name of the gene to be plotted.
}
\examples{
scubi_expression(dim1=rnorm(10000),dim2=rnorm(10000),count=matrix(rbinom(100000,10,0.1),nrow=10,dimnames=list(LETTERS[1:10],NULL)),gene='A')
scubi_expression(dim1=rnorm(10000),dim2=rnorm(10000),count=matrix(rbinom(100000,10,0.1),nrow=10,dimnames=list(LETTERS[1:10],NULL)),gene=list('A','B'))
}
\author{
Wenpin Hou<whou10@jhu.edu>, Zhicheng Ji
}
