% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scubi_sample_composition.R
\name{scubi_sample_composition}
\alias{scubi_sample_composition}
\title{SCUBI for multi-sample cell type compositions comparison}
\usage{
scubi_sample_composition(
  dim1,
  dim2,
  sample,
  design,
  smooth = TRUE,
  clip = TRUE,
  scale = TRUE,
  resolution = 10 * 20/max(diff(range(dim1)), diff(range(dim2))),
  plot = TRUE,
  palette = rainbow(15)[c(11:1, 15)]
)
}
\arguments{
\item{dim1}{Numeric vector for the first dimension, such as the first dimension of UMAP or PCA.}

\item{dim2}{Numeric vector for the second dimension.}

\item{sample}{Character vector of cell-sample assignment (each cell belong to which sample).}

\item{design}{A numeric data.frame or matrix indicating the covariates of the samples. It should not contain the intercept term. The row names of the matrix should contain the unique values in 'sample' variable.}

\item{smooth}{Either TRUE or FALSE. If TRUE, the coefficients will be smoothed across the coordinate space with loess.}

\item{clip}{Either TRUE or FALSE. If TRUE, will enforce values larger than 0.975 quantile to equal 0.975 quantile, and enforce values smaller than 0.025 quantile to equal 0.025 quantile.}

\item{scale}{Either TRUE or FALSE. If TRUE, the values will be scaled to have mean 0 and standard deviation of 1.}

\item{resolution}{Numeric value of resolution. A higher value will result in smaller squares, thus higher resolution.}

\item{plot}{Either TRUE or FALSE. If TRUE, the function will directly generate a plot. If FALSE, the function will return the data frame used for plotting. This is useful for more complicated operations.}

\item{palette}{A vector of color codes for plotting palette.}
}
\value{
If plot=TRUE, a ggplot object. If plot=FALSE, a data.frame with the following columns: xmin,xmax,ymin,ymax of the square, and the averaged expresison value of that square.
}
\description{
SCUBI for comparing cell type compositions across samples
}
\details{
This function partitions the whole coordinate space into multiple non-overlapping squares, calculates the proportions of discrete categories for cells falling in each square, and calculates and visualizes the differences between sample categories. It takes as input the coordinates of the cells (e.g. UMAP or PCA), the cell-sample assignment (each cell belong to which sample), and a numeric design matrix for the samples. An intercept will be automatically added to the design matrix so do not include any intercept term. The first column of the design matrix will always be the contrast of interest and will be visualized.
}
\examples{
scubi_sample_composition(dim1=rnorm(10000),dim2=rnorm(10000),sample=sample(c('A','B'),10000,replace=TRUE),design=data.frame(group=c(1,0),row.names=c('A','B')))
}
\author{
Wenpin Hou<whou10@jhu.edu>, Zhicheng Ji
}
