% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_BBS_model.R
\name{setup_BBS_model}
\alias{setup_BBS_model}
\title{Create input data for fitting BBS-style hierarchical model for estimating trends in point count survey data}
\usage{
setup_BBS_model(dat, count = "count_max")
}
\arguments{
\item{dat}{Dataframe or tibble containing point count data}

\item{count}{Character string giving the name of the field containing the count data to be modeled.}
}
\value{
A list containing all inputs necessary for fitting BBS-style hierarchical model.
}
\description{
Create input data for fitting BBS-style hierarchical model for estimating trends in point count survey data
}
