% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_by.R
\name{summarize_by}
\alias{summarize_by}
\title{Summarize banding statistics (internal function)}
\usage{
summarize_by(
  df,
  by = "month",
  stat = c("nethours", "captures"),
  winter_adjust = NULL
)
}
\arguments{
\item{df}{Input dataframe}

\item{by}{how statistics should be summarized, either "month", "year", or
"season"}

\item{stat}{the type of stats to be summarized; either "nethours" or
"captures"}

\item{winter_adjust}{logical}
}
\value{
tibble
}
\description{
Internal functions called by [summarize_capture_stats()] and
[summarize_nethrs_stats()] and not intended to be called independently.
}
\details{
See documentation for [summarize_capture_stats()] for details. Note
that the only valid options for the argument `stat` are currently: "nethours"
or "captures". This value determines whether the NETHOURS field is summed or
the number of BANDNUMBs are counted, and the name of the column containing
the results.
}
\keyword{internal}
