% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_banding_stats.R,
%   R/summarize_nethrs_stats.R
\name{summarize_capture_stats}
\alias{summarize_capture_stats}
\alias{summarize_nethrs_stats}
\title{Summarize banding statistics}
\usage{
summarize_capture_stats(
  path = NULL,
  df = NULL,
  location = NULL,
  datemin = NULL,
  datemax = NULL,
  species = NULL,
  by = "year",
  winter_adjust = TRUE
)

summarize_nethrs_stats(
  path = NULL,
  df = NULL,
  location = NULL,
  datemin = NULL,
  datemax = NULL,
  by = "year",
  winter_adjust = TRUE
)
}
\arguments{
\item{path}{optional filepath to local copy of banding or net hours database}

\item{df}{optional data frame representing banding or net hours database}

\item{location}{optional character string of banding station codes to limit
results}

\item{datemin}{optional minimum date in YYYY-mm-dd format}

\item{datemax}{optional maximum date in YYYY-mm-dd format}

\item{species}{optional character string of species codes to include (only
applies to banding data)}

\item{by}{how statistics should be summarized, either "month", "year", or
"season"; see Details}

\item{winter_adjust}{logical; see Details}
}
\value{
tibble
}
\description{
`summarize_capture_stats()` returns total captures by location,
species, and either month, year, or season

`summarize_nethrs_stats()` returns total net hours by location, and either
month, year or season
}
\details{
These functions are designed to work directly with standard
  Palomarin .dbf files, e.g. "allnumb.dbf", "band.dbf", or
  "allpalonthrs.dbf". If `df` is provided, it should represent a local
  dataframe version of the banding or net hours database already in memory,
  potentially already filtered to the desired locations, date range, and/or
  species, or with a custom "season" column added. If `df` is not provided,
  instead use `path` to point to a .dbf file that will be read in using
  [foreign::read.dbf()].

  The `location`, `datemin`, `datemax`, and `species` arguments are optional
  and provide convenient ways to filter the data as needed, especially if
  reading in the .dbf directly from the `path` provided. Note that `species`
  filters apply only to the banding data.

  After filtering, the internal [summarize_by()] function is called to either
  count the number of captures or sum the number of net hours as appropriate,
  summarized by location, species (for banding data only), and the desired
  time frame specified using the `by` argument. Currently supported values
  include "month", "season", or "year". If `by = "month"`, capture statistics
  will be calculated by location, species (for banding data only), year, and
  month. If `by = "season"`, summary statistics will be calculated by
  location, species (for banding data only), year, and season. If `by =
  "year"`, capture statistics will be calculated by location, species (for
  banding data only), and year.

  If the corresponding fields "month", "season", or "year" are already
  present in the `df` provided, those are the fields that will be used to
  calculate these summary statistics, and thereby provide a means to
  customize how month, year, or season are defined. If those fields are not
  present in `df`, or if `path` is provided, the corresponding month and year
  will be calculated from the "DATE" field, and default seasons will be
  defined as: spring (March through July), fall (August through October), and
  winter (November through February). Note that in this case, there must be
  "DATE" field present in `df` or the database to which `path` points.

  If `by = 'season'`, with `winter_adjust = TRUE` (the default), there will
  be an additional field created called "year_adjust". This field will be
  identical to "year" except that when `season == "winter"` and `month < 6`,
  year_adjust will be equal to year - 1, so that, for example, January will
  be grouped with December from the previous calendar year. In this case,
  summary statistics will be calculated by location, species (for banding
  data only), and year_adjust.

  Finally, the resulting dataframe will be expanded to sure there is a unique
  row for every location, year, month, season, and/or species, including
  where and when there may have been zero captures or zero net hours in the
  data.
}
\examples{

data(sample_band)
capture_MUHO = summarize_capture_stats(
   df = sample_band, location = 'MUHO', species = c('SWTH', 'AMGO'),
   datemin ='2010-01-01', by = 'season', winter_adjust = FALSE)

data(sample_nethrs)
effort_PGUP = summarize_nethrs_stats(
   df = sample_nethrs, location = 'PGUP', datemax ='2010-10-15',
   by = 'month') 
\dontrun{
# # summarize seasonal capture statistics for SOSP and WIWA directly from
}
}
