% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_distance_key.R
\name{add_distance_key}
\alias{add_distance_key}
\title{Convert Distance Bins to numeric mindist and maxdist fields}
\usage{
add_distance_key(dat)
}
\arguments{
\item{dat}{Dataframe containing the field 'Distance Bin'}
}
\value{
Dataframe
}
\description{
Convert Distance Bins to numeric mindist and maxdist fields
}
