% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_PC_dat.R
\name{compile_PC_dat}
\alias{compile_PC_dat}
\title{Compile point count survey data from a directory}
\usage{
compile_PC_dat(dir, pattern = ".csv")
}
\arguments{
\item{dir}{Directory containing point count survey data downloaded from CADC}

\item{pattern}{Regex pattern for identifying files to include (see \code{\link[base]{list.files}})}
}
\value{
Dataframe of compiled data, excluding any rows with all NA values.
}
\description{
Compile point count survey data from a directory
}
\details{
Also tests for missing values in critical fields.
}
