% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BBS_model.R
\name{fit_BBS_model}
\alias{fit_BBS_model}
\title{Fit BBS-style hierarchical model for estimating trends in point count survey data}
\usage{
fit_BBS_model(inputdata, n.adapt = 500, n.update = 500,
  n.iter = 1000, n.chains = 3, overdispersion = TRUE,
  randomslope = FALSE, ...)
}
\arguments{
\item{inputdata}{inputdata created by running \code{\link{setup_BBS_model}}}

\item{n.adapt}{number of iterations for adaptation, defaults to 500}

\item{n.update}{number of iterations for burn-in, defaults to 500}

\item{n.iter}{number of iterations for sampling, defauts to 1000}

\item{n.chains}{number of MCMC chains}

\item{overdispersion}{defaults to TRUE, to include a term in the model for
fitting overdispersion, as in BBS models}

\item{randomslope}{defaults to FALSE, because this is not included in BBS
models; change to TRUE to include a correlated random slope by transect, in
addition to a random intercept}
}
\value{
Returns a coda.samples object from rjags. Also displays summary
statistics from MCMCsummary and prints trace plots from MCMCplot.
}
\description{
Fit BBS-style hierarchical model for estimating trends in point count survey data
}
