% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BBS_model_estimates.R
\name{get_BBS_model_estimates}
\alias{get_BBS_model_estimates}
\title{Extract annual trend or abundance index estimates from BBS-style hierarchical model}
\usage{
get_BBS_model_estimates(modresults, inputdat, type)
}
\arguments{
\item{modresults}{Model output from running \code{\link{fit_BBS_model}}}

\item{inputdat}{Input data from running \code{\link{setup_BBS_model}}}

\item{type}{Type of estimates to extract: annual indices of abundance
('index'), or predicted values from the linear trend ('trend').}
}
\value{
Dataframe containing predicted value and HDI for each year present
in inputdat$year.pred (created by \code{\link{setup_BBS_model}})
}
\description{
Extract annual trend or abundance index estimates from BBS-style hierarchical model
}
\details{
This function is automatically called by \code{\link{plot_BBS_model}},
but may be useful for generating customized plots.
}
