% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palo.R
\name{Palo}
\alias{Palo}
\title{Palo: Color palette optimization}
\usage{
Palo(position, cluster, palette, iter = 1000)
}
\arguments{
\item{position}{A matrix of low-dimensional coordinates for single-cell data or 2-D spatial positions for spatial transcriptomics data. The number of rows of the matrix is the number of cells and the matrix has two columns.}

\item{cluster}{A vector of cell or spots clusters. Should have the same length as the number of cells or spots. Should have the same order as the columns of \code{position}.}

\item{palette}{A user-defined character vector of palette.}

\item{iter}{A numeric value of optimization iterations.}
}
\value{
A named vector of colors.
}
\description{
Palo: Color palette optimization for single-cell and spatial data
}
\examples{
Palo(matrix(rnorm(200),ncol=2),rep(c('c1','c2'),50),c('red','blue'))
}
\author{
Wenpin Hou<whou10@jhu.edu>, Zhicheng Ji
}
