% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BBS_model.R
\name{plot_BBS_model}
\alias{plot_BBS_model}
\title{Plot trend results from BBS-style hierarchical model for point count survey data}
\usage{
plot_BBS_model(modresults, inputdat, type = "annual+smooth_overall")
}
\arguments{
\item{modresults}{model output from running \code{\link[MASS]{fit_BBS_model}}}

\item{inputdat}{input data from running \code{\link[MASS]{setup_BBS_model}}}

\item{type}{type of plot to produce: plots of annual indices of abundance
overall or by transect ('annual_overall', 'annual_transect'), plots of
linear trend lines overall or by transect ('smooth_overall',
'smooth_transect'), or combinations of both ('annual+smooth_overall',
'annual+smooth_transect'). Defaults to 'annual+smooth_overall', similar to
BBS output.}

\item{plot}{defaults to TRUE, to display a plot of results}
}
\value{
returns ggplot
}
\description{
Plot trend results from BBS-style hierarchical model for point count survey data
}
