% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BBS_model.R
\name{plot_BBS_model}
\alias{plot_BBS_model}
\title{Plot trend results from BBS-style hierarchical model for point count survey data}
\usage{
plot_BBS_model(modresults, inputdat, type = "global_trend+index")
}
\arguments{
\item{modresults}{Model output from running \code{\link{fit_BBS_model}}}

\item{inputdat}{Input data from running \code{\link{setup_BBS_model}}}

\item{type}{Type of plot to produce: plots of annual indices of abundance
overall or by transect ('global_index', 'transect_index'), plots of
linear trend lines overall or by transect ('global_trend',
'transect_trend'), or combinations of both ('global_trend+index',
'transect_trend+index'). Defaults to 'global_trend+index', similar to
BBS output.}
}
\value{
returns ggplot
}
\description{
Plot trend results from BBS-style hierarchical model for point count survey data
}
\details{
Runs \code{\link{get_BBS_model_estimates}} and produces ggplot
}
