% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scubi_discrete.R
\name{scubi_discrete}
\alias{scubi_discrete}
\title{SCUBI for discrete variable}
\usage{
scubi_discrete(
  dim1,
  dim2,
  feature,
  normalize = FALSE,
  resolution = 10 * 20/max(diff(range(dim1)), diff(range(dim2))),
  plot = TRUE,
  palette = rainbow(length(unique(feature)))
)
}
\arguments{
\item{dim1}{Numeric vector for the first dimension, such as the first dimension of UMAP or PCA.}

\item{dim2}{Numeric vector for the second dimension.}

\item{feature}{Character vector of a discrete variable (e.g. gender of the sample the cell was collected).}

\item{normalize}{Either TRUE or FALSE. If TRUE, for each discrete category, number of cells in each region will be normalized by the total number of cells. This is to address the bias induced by the unbalanced total number of cells for discrete categories.}

\item{resolution}{Numeric value of resolution. A higher value will result in smaller squares, thus higher resolution.}

\item{plot}{Either TRUE or FALSE. If TRUE, the function will directly generate a plot. If FALSE, the function will return the data frame used for plotting. This is useful for more complicated operations.}

\item{palette}{A vector of color codes for plotting palette.}
}
\value{
If plot=TRUE, a ggplot object. If plot=FALSE, a data.frame with the following columns: xmin,xmax,ymin,ymax of the square, and the averaged expresison value of that square.
}
\description{
SCUBI for plotting a discrete variable
}
\details{
This function partitions the whole coordinate space into multiple non-overlapping squares, and calculates and visualizes the proportions of discrete categories for cells falling in each square. It takes as input the coordinates of the cells (e.g. UMAP or PCA) and the values of the discrete variable.
}
\examples{
scubi_discrete(dim1=rnorm(10000),dim2=rnorm(10000),feature=sample(c('A','B'),10000,replace=TRUE))
}
\author{
Wenpin Hou<whou10@jhu.edu>, Zhicheng Ji
}
