% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_PC_dat.R
\name{summarize_PC_dat}
\alias{summarize_PC_dat}
\title{Format and summarize point count survey data for a given species and project code.}
\usage{
summarize_PC_dat(df, species, project)
}
\arguments{
\item{df}{master point count data set, with dates, transects, and distances filtered}

\item{species}{4-letter species code}

\item{project}{4-letter project code}
}
\value{
dataframe containing all Points, Transects, and Years in the data,
the number of visits, the max and mean count for the species selected,
including zero counts where the species was not detected, and the standard
error of the mean count
}
\description{
Format and summarize point count survey data for a given species and project code.
}
