% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_PC_dat.R
\name{summarize_PC_dat}
\alias{summarize_PC_dat}
\title{Format and summarize point count survey data for a given species and project code.}
\usage{
summarize_PC_dat(df, species, project)
}
\arguments{
\item{df}{Data frame of point count data, with the standard field names as
in the format downloaded from CADC Project Leader. Filter to include only
the detection distances that should be included in the total count.}

\item{species}{Character string for the 4-letter species code, in all caps}

\item{project}{Character string for the 4-letter project code, in all caps}
}
\value{
Dataframe containing the total count (regardless of detection
distance) for the selected species at each unique combination of Visit,
Point, Transect, and Year in the data, including zero counts where the
species was not detected.
}
\description{
Format and summarize point count survey data for a given species and project code.
}
